Rework on handling of sanitizer server without using Emissary API's.
