+++ clang/lib/Driver/ToolChain.cpp
@@ -1681,7 +1681,7 @@ SanitizerMask ToolChain::getSupportedSanitizers() const {
     Res |= SanitizerKind::MemTag;
   if (getTriple().isBPF())
     Res |= SanitizerKind::KernelAddress;
-  if (getTriple().isAMDGPU())
+  if(getTriple().isAMDGPU())
     Res |= SanitizerKind::Address;
   return Res;
 }
