+++ offload/plugins-nextgen/amdgpu/src/rtl.cpp
@@ -4884,6 +4884,9 @@ struct AMDGPUDeviceTy : public GenericDeviceTy, AMDGenericDeviceTy {
 
   bool useStrictSanityChecks() const { return OMPX_StrictSanityChecks; }
 
+  /// Get the device memory heap pointer for ASan.
+  void *getDMHeapPtr() const { return DMHeapPtr; }
+
 private:
   using AMDGPUEventRef = AMDGPUResourceRef<AMDGPUEventTy>;
   using AMDGPUEventManagerTy = GenericDeviceResourceManagerTy<AMDGPUEventRef>;
@@ -6092,6 +6095,10 @@ Error AMDGPUKernelTy::launchImpl(GenericDeviceTy &GenericDevice,
 
     hsa_utils::initImplArg(ImplArgs, &ImplArgsTy::DynamicLdsSize, ImplArgsSize,
                            KernelArgs.DynCGroupMem);
+
+    hsa_utils::initImplArg(
+        ImplArgs, &ImplArgsTy::HeapV1Ptr, ImplArgsSize,
+        reinterpret_cast<uint64_t>(AMDGPUDevice.getDMHeapPtr()));
   }
 
   // Get required OMPT-related data
